Refactor Inventory location
Move the class to its own file, makes working on Stream easier.
